services/burvesigner/src/requirements.txt
fastecdsa
Flask
pycryptodome